requirements.txt
Flask==3.0.0
openai==0.27.2
python-dotenv==0.21.0